pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-news-scheduler"
version = "0.1.0"
description = "Celery Beat service that triggers the daily content pipeline for each configured topic"
requires-python = ">=3.11"
dependencies = [
    "celery[redis]>=5.3",
    "httpx>=0.27",
    "pydantic-settings>=2.2",
    "python-slugify>=8.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "anyio>=4.3",
    # db-service installed as local package via Docker COPY
    "ai-news-db @ file:///db-service",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
